Construct Initial Hierarchy

This reshapes the list output from node_totals into the hierarchical format needed for the d3 tree visualization.

Usage

node_hierarchy(tree, totals, node = NULL)

Arguments

tree

An object of class phylo, representing the tree structure.

totals

A named list of node totals, as returned by node_totals.

node

Name of the node from which to start a recursion. Defaults to the root node.

Value

A nested list representing the hierarchy, with each node containing ' its name, value, summary, and children (if any).

Examples

library(ape)
set.seed(1)
tree <- rtree(5)
x_mat <- matrix(runif(15), ncol = 5)
colnames(x_mat) <- tree$tip.label

totals <- node_totals(tree, x_mat)
node_hierarchy(tree, totals)
#> Error in map(children_nodes, ~node_hierarchy(tree, totals, .x)): ℹ In index: 1.
#> Caused by error in `matrix()`:
#> ! 'data' must be of a vector type, was 'NULL'